refactor!: move optimizations hash set to ImageHandler fixes #52, add new struct "ImageModifications" as `ImageOptimizations` is separating into a UI / config front facing enum, switch to `ImageModifications` and split ``ImageHandler`` up into folder module. #25
