+++ src/image/mod.rs
@@ -1,6 +1,6 @@
 pub mod image;
-pub mod optimization;
-pub mod image_formats;
 pub mod backends;
+pub mod image_formats;
+pub mod modifications;
 
 mod fast_downsample;
