+++ src/image_handler/dynamic_sampling.rs
@@ -3,15 +3,16 @@ use std::time::Duration;
 use eframe::egui::Vec2;
 use log::debug;
 
-use crate::{image::optimization::{get_monitor_downsampling_size, EventImageOptimizations, ImageOptimizations, InitialImageOptimizations}, image_handler::ImageHandler, monitor_size::MonitorSize, scheduler::Scheduler, zoom_pan::ZoomPan};
+use crate::{image::image::ImageSizeT, image_handler::{monitor_downsampling::get_monitor_downsampling_size, optimization::ImageOptimizations}, monitor_size::MonitorSize, scheduler::Scheduler, zoom_pan::ZoomPan};
+
+use super::ImageHandler;
 
 impl ImageHandler {
     pub fn dynamic_sampling_update(&mut self, zoom_pan: &ZoomPan, monitor_size: &MonitorSize) {
         if let Some(image) = &self.image {
-            // the zoom factor change since the last dynamic upsample / downsample.
-            let is_enabled = image.optimizations.contains(
-                &ImageOptimizations::EventBased(EventImageOptimizations::DynamicUpsampling)
-            );
+            let is_enabled = self.has_optimization(
+                &ImageOptimizations::DynamicSampling(bool::default(), bool::default())
+            ).is_some();
 
             if zoom_pan.zoom_factor <= 1.0 || !is_enabled {
                 self.last_zoom_factor = 1.0;
@@ -34,12 +35,12 @@ impl ImageHandler {
             // TODO: we seriously need to find about setting image size universally to some type 
             // like (usize, usize), (u32, u32) or (f32, f32) because those "as" statements below is not it bro.
 
-            // TODO: if monitor downsampling is set use monitor downsampled image size instead of normal image size.
             let max_image_size = (image.image_size.width as u32, image.image_size.height as u32);
             let mut image_size = max_image_size;
 
-            if let Some(ImageOptimizations::Initial(InitialImageOptimizations::MonitorDownsampling(marginal_allowance))) = image.has_optimization(
-                &ImageOptimizations::Initial(InitialImageOptimizations::MonitorDownsampling(0))
+            // TODO: (28/03/2025) check if we even need this now
+            if let Some(ImageOptimizations::MonitorDownsampling(marginal_allowance)) = self.has_optimization(
+                &ImageOptimizations::MonitorDownsampling(u32::default())
             ) {
                 image_size = get_monitor_downsampling_size(*marginal_allowance, image_size, monitor_size);
             }
@@ -48,11 +49,10 @@ impl ImageHandler {
                 Vec2::new(image_size.0 as f32, image_size.1 as f32)
             );
             let new_resolution = (
-                new_resolution.x.clamp(image_size.0 as f32, max_image_size.0 as f32),
-                new_resolution.y.clamp(image_size.1 as f32, max_image_size.1 as f32)
+                new_resolution.x.clamp(0.0, max_image_size.0 as f32) as u32,
+                new_resolution.y.clamp(0.0, max_image_size.1 as f32) as u32
             );
 
-            // TODO: also don't upsample if image is already at it's max resolution.
             if self.accumulated_zoom_factor_change >= change {
                 // TODO: schedule a image reload with new dimensions.
 
@@ -69,27 +69,30 @@ impl ImageHandler {
         }
     }
 
-    pub fn schedule_image_dynamic_sample(&mut self, upsample: bool, resolution: (f32, f32)) {
-        if resolution == self.dynamic_sampling_old_resolution {
-            debug!(
-                "Not scheduling dynamic image sample for '{:.0}x{:.0}' \
-                as that resolution has already been set or scheduled.",
-                resolution.0,
-                resolution.1
-            );
-            return;
-        }
+    pub fn schedule_image_dynamic_sample(&mut self, upsample: bool, resolution: ImageSizeT) {
+        // if resolution == self.dynamic_sampling_old_resolution {
+        //     debug!(
+        //         "Not scheduling dynamic image sample for '{:.0}x{:.0}' \
+        //         as that resolution has already been set or scheduled.",
+        //         resolution.0,
+        //         resolution.1
+        //     );
+        //     return;
+        // }
 
         let delay = match upsample {
             true => Duration::from_secs_f32(1.5),
             false => Duration::from_secs(5),
         };
 
+        // TODO: fix lifetime reference error here
         let schedule = Scheduler::new(
             move || {
-                // TODO: figure out how to tell the Image or ImageLoader struct 
-                // to reload the image with DownSampling or UpSampling optimizations.
-                println!("IMAGE SAMPLED! {:?}", resolution);
+                self.dynamic_sampling_new_resolution = resolution;
+                // TODO: implement reloading mechanician into 
+                // ImageHandler::load_image then we are safe to call it here.
+
+                // self.load_image();
             },
             delay
         );
