+++ src/main.rs
@@ -30,7 +30,6 @@ mod window_scaling;
 mod magnification_panel;
 mod monitor_size;
 mod scheduler;
-mod dynamic_sampling;
 
 /// 🌹 A fast as fuck, memory efficient and simple but fancy image viewer built with 🦀 Rust that's cross platform.
 #[derive(Parser, Debug)]
@@ -127,7 +126,7 @@ fn main() -> eframe::Result {
                 error.into(), ToastLevel::Error
             ).duration(Some(Duration::from_secs(10)));
         } else {
-            let result = image_handler.init_image(path, &monitor_size);
+            let result = image_handler.init_image(path);
 
             if let Err(error) = result {
                 notifier.toasts.lock().unwrap().toast_and_log(
